requirements.txt
python-telegram-bot==21.6
fastapi==0.115.0
uvicorn==0.34.0
pydantic-settings==2.7.1
sqlalchemy[asyncio]==2.0.46
asyncpg==0.30.0
